Flytt nasjonal-utils til tilhørende utils-fil
