+++ src/schemas/ba-sak/begrunnelse/nasjonaleTriggere/endretUtbetalingPeriodeTrigger.ts
@@ -1,8 +1,7 @@
 import { BegrunnelseDokumentNavn, SanityTyper } from '../../../../util/typer';
 import { endretUtbetalingsperioderTriggereValg } from '../typer';
 import { erEndretUtbetaling } from './endringsårsakTrigger';
-import { hentNasjonaleTriggereRegler } from './utils';
-import { erNasjonalBegrunnelse } from '../utils';
+import { hentNasjonaleTriggereRegler, erNasjonalBegrunnelse } from './utils';
 
 export const endretUtbetalingsperiodeTriggere = {
   title: 'Endret utbetalingsperiode triggere',
