+++ src/schemas/ba-sak/begrunnelse/nasjonaleTriggere/utvidetBarnetrygdTriggere.ts
@@ -1,8 +1,7 @@
 import { BegrunnelseDokumentNavn, SanityTyper } from '../../../../util/typer';
 import { utvidetBarnetrygdTriggertyper, Vilkår, vilkårTriggerTilMenynavn } from '../typer';
 import { erEøsBegrunnelse } from '../eøs/eøsTriggere/utils';
-import { hentNasjonaleTriggereRegler } from './utils';
-import { erNasjonalBegrunnelse } from '../utils';
+import { hentNasjonaleTriggereRegler, erNasjonalBegrunnelse } from './utils';
 
 export const utvidetBarnetrygdTriggere = {
   title: 'Utvidet barnetrygd triggere',
